# ta-api-master/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/dump/geography/countries": {
      "get": {
        "operationId": "GeographyDumpController_getCountryList",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "countryName",
            "required": false,
            "in": "query",
            "description": "Country Name",
            "schema": {
              "example": "India",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get country list",
        "tags": [
          "Geography Dump"
        ]
      }
    },
    "/dump/geography/states": {
      "get": {
        "operationId": "GeographyDumpController_getStateList",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "countryCode",
            "required": true,
            "in": "query",
            "description": "Country Code",
            "schema": {
              "example": "IN, US, etc.",
              "type": "string"
            }
          },
          {
            "name": "stateName",
            "required": false,
            "in": "query",
            "description": "State Name",
            "schema": {
              "example": "Maharashtra",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get state list",
        "tags": [
          "Geography Dump"
        ]
      }
    },
    "/dump/geography/cities": {
      "get": {
        "operationId": "GeographyDumpController_getCityList",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stateCode",
            "required": true,
            "in": "query",
            "description": "State Code",
            "schema": {
              "example": "BDS",
              "type": "string"
            }
          },
          {
            "name": "countryCode",
            "required": true,
            "in": "query",
            "description": "Country Code",
            "schema": {
              "example": "IN, US, etc.",
              "type": "string"
            }
          },
          {
            "name": "cityName",
            "required": false,
            "in": "query",
            "description": "City Name",
            "schema": {
              "example": "Mumbai",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get city list",
        "tags": [
          "Geography Dump"
        ]
      }
    },
    "/dump/geography/update-city-vector": {
      "get": {
        "operationId": "GeographyDumpController_updateCityVectorAndCityNameNormalized",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update city vector and city name normalized",
        "tags": [
          "Geography Dump"
        ]
      }
    },
...
  },
  "info": {
    "title": "API SERVICE",
    "description": "## Travel API Documentation\n\nWelcome to the **API SERVICE** documentation. This guide provides comprehensive guidelines and detailed instructions on how to effectively integrate and utilize the Travel API.\n\n### Overview\nThis documentation includes various sections covering:\n- **Authentication Methods**: Learn how to authenticate using Bearer and Cookie authentication.\n- **Endpoint Descriptions**: Detailed information on each API endpoint.\n- **Usage Examples**: Practical examples to help you get started quickly.\n- **Best Practices**: Tips and recommendations for efficient implementation.\n\n### Purpose\nWhether you are a developer looking to incorporate travel-related functionalities or a business aiming to enhance your travel services, this documentation serves as a valuable resource to guide you through the process.\n\n### Getting Started\nTo get started, ensure you have the necessary authentication tokens and follow the endpoint descriptions to make your first API call.\n\nFor any questions or support, please refer to the contact section at the end of this documentation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
...
  }
}